Replace the R128 meters for each channel with a digital peak meter.
R128 is a superior way of measuring loudness, but for each bus, we care more
about getting the recording levels right (ie., that we don't peak) than loudness.
Since we do our mixing in float, unlike a typical digital mixer, it's not as
critical (at least assuming that the limiter is on and will save us later in
the mastering chain), but it's probably better to be consistent with convention
here.

Thus, use a simple digital meter for each bus, even though we of course still
use R128 for the mastering section.