Add mute buttons.
Actually this is mostly because I like being able to press the
mute button on my MIDI controller and have it light up :-)
Internally, it's implemented by forcing the volume to zero
(or sometimes -90 dB), which means we get all the nice
fading properties and such.

In the process, we now declare -90 dB to be truly zero, and don't
bother adding it to the bus.