Fix an issue with the correction factor locking to 0.95.
Basically what was happening is that if the master card lost
or corrupted a frame, which we didn't set a timestamp on,
causing it to have steady_clock::time_point::min(). This would
in turn cause us to assume a latency of trillions of seconds,
throwing off the filter and essentially making it be 0.95 forever.

The fix is twofold; we always set timestamps, but also make
ourselves robust to the ones that are way off (negative uptime).