Rework the queue drop algorithm again.
The old one worked fine for most cases, but when two cards drift
very slowly out of sync (and one of them is the master), it's inadequate;
we should simply drop one frame at the right moment and be done with it.
Of course, knowing what the right moment is can be tricky, due to jitter,
which we now explicitly model.

Also include a separate program that tests various properties of different
possible policies (used to determine which one to use), based on replaying
real event logs.